+++ CHANGELOG.md
@@ -65,6 +65,9 @@
 - `BrowserCluster`
     - `Worker`
         - `#browser_respawn` -- Catch Watir/Selenium errors.
+- `Check`
+    - `Auditor`
+        - `#each_candidate_dom_element` -- Yield element DOMs instead of parent elements.
 -  Checks
     - Active
         - Added